+++ tf/advanced/app.py
@@ -351,7 +351,7 @@ def findApp(
 
     host: string, optional None
         If present, it points to a GitLab instance such as the on-premise
-        `gitlab.huc.knaw.nl`.
+        `gitlab.huc.knaw.nl` or the public `gitlab.com`.
         If `None` we work with `github.com`.
 
     kwargs: mixed