+++ tf/advanced/data.py
@@ -18,7 +18,7 @@ class AppData(object):
         ----------
         host: string, optional None
             If present, it points to a GitLab instance such as the on-premise
-            `gitlab.huc.knaw.nl`.
+            `gitlab.huc.knaw.nl` or the public `gitlab.com`.
             If `None` we work with `github.com`.
         app: obj
             The high-level API object